resources/multi-account-manager/requirements.txt
# Multi-Account Strategy Manager — Dependencies
# Install: pip install -r requirements.txt

# Alpaca SDKs (both needed: old for FVG strategy, new for manager + tradingView)
alpaca-py>=0.21.0
alpaca-trade-api>=3.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
dash>=2.14.0
plotly>=5.18.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment
python-dotenv>=1.0.0

# Timezone support
pytz>=2023.3

# Crypto exchange (needed by candlestick_pro)
ccxt>=4.0.0
